Add uworker and utils
Summary:
First version of uworker. The plan is to make a docker image that installs the
worker and then build on that image to add job specific installations. The worker container is then started with the proper environment vars.

The worker needs some new features from the api:
- Support for project name
- Job output url
- Job type

Test Plan: test/test_uworker.py

Reviewers: joakim, skymandr

Reviewed By: joakim, skymandr

Differential Revision: http://phabricator.molflow.com/D40