Adapt uworker to running on host
Summary:
If the worker is running on the host it will fetch jobs from the project prio endpoint
and assumes that the projects provide docker image urls for processing the jobs.

The docker images are downloaded (docker pull) and the jobs executed via "docker run".

The worker can still run in docker as before, these changes are backwards-compatible.

This diff also includes build scripts for building a virtualbox image and setup a test environment
to be able to run the worker in a local vm from local code.

Test Plan: arc unit

Reviewers: skymandr, joakim

Reviewed By: skymandr, joakim

Differential Revision: http://phabricator.molflow.com/D129