Changeing the configuration of the connectionpool with hope that this will work better in the production environment.
Summary:
A "almost done review" - I'm not conviced that this works.
The connection pool was set up with standard values, I've decreased the recycle time (to release frozen connection earlier)
and increased the timeout (to let clients wait a little bit longer for a connection) set the max_overflow to 0 and increased
pool size.

I also fixed a type in the code - but I'm not 100% sure that it changes the behaviour of the database. It might change the instatiantion of a db_session

Test Plan:
Deploy and se if it works better?
I've added a two tests with conncurrent lookups (but I'm not sure that it works correctly)

Reviewers: skymandr, gregoire

Reviewed By: skymandr, gregoire

Subscribers: jenkins

Maniphest Tasks: T562

Differential Revision: https://phabricator.molflow.com/D327