Allow posting a list of jobs
Summary:
If the top-level json object that is recieved is a list, all the jobs will be
added. If it's a dictionary, it'll treat it as a single job (so it should be
backward compatible).

In case one or more job can't be added (invalid, duplicate id, etc) no job is
added.

Test Plan: Add tests for the different cases I could think of.

Reviewers: joakim, bengt.rydberg, skymandr

Reviewed By: joakim, skymandr

Subscribers: jenkins

Maniphest Tasks: T515

Differential Revision: https://phabricator.molflow.com/D330